% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Calculates the Euclidean distance between simulated data and observed data}
\usage{
euclidean_distance(sim_data, obs_data)
}
\arguments{
\item{sim_data}{Simulated data}

\item{obs_data}{Observed data}
}
\value{
Euclidean distance
}
\description{
Calculates the Euclidean distance between simulated data and observed data
}
