% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{general_sample_data}
\alias{general_sample_data}
\title{Generates a vector of i.i.d. random sums}
\usage{
general_sample_data(params, n)
}
\arguments{
\item{params}{List of distributional arguments for the sum and summand distributions}

\item{n}{Number of samples to generate}
}
\value{
Vector of i.i.d. random sums
}
\description{
Generates a vector of i.i.d. random sums
}
