% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc.R
\name{smc}
\alias{smc}
\title{Sequential Monte Carlo algorithm}
\usage{
smc(
  n_particles,
  priors,
  simulate_data,
  obs_data,
  distance_function,
  epsilon = Inf,
  n_iter = 1
)
}
\arguments{
\item{n_particles}{Number of particles to sample}

\item{priors}{List of prior distributions}

\item{simulate_data}{Function to simulate data from parameters}

\item{obs_data}{Observed data}

\item{distance_function}{Function to calculate the distance between the fake data and the observed data}

\item{epsilon}{Maximum distance for a particle to be accepted}

\item{n_iter}{Number of iterations}
}
\value{
List of populations
}
\description{
Sequential Monte Carlo algorithm
}
