% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{wasserstein_distance}
\alias{wasserstein_distance}
\title{Calculates the empirical Wasserstein distance between simulated data and observed data}
\usage{
wasserstein_distance(sim_data, obs_data)
}
\arguments{
\item{sim_data}{Simulated data}

\item{obs_data}{Observed data}
}
\value{
Empirical Wasserstein distance
}
\description{
Calculates the empirical Wasserstein distance between simulated data and observed data
}
